% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{fetch_ts_batched}
\alias{fetch_ts_batched}
\title{Helper Function to batch reading ts}
\usage{
fetch_ts_batched(keys, batch_size = 50)
}
\arguments{
\item{keys}{Time series Keys}

\item{batch_size}{how many keys to read from db at once}
}
\description{
Helps batch the keys to fetch from timeseries db
}
